moved test until utils with other dns tests, split tuple into two values
